Allow reification of third-party libraries, closes #14
We configure the overriding with a `typeMap: Map[Type, Type]`.  We
perform the type overriding as an integral part of the ReifyAsEmbedding
phase, more precisely in the `body` method.

The benefit of this approach is that we don't have to type-rewrite any
nodes in the block. This means the reification of standard and
third-party types will only trigger if the DSL author provides an
implementation of the invoked method in the corresponding type in
`typeMap`. If no matching method (see bottom comment) with a `reifyAs`
annotation is found in the corresponding type the transformation falls
back to the general error message, i.e. 'method X is not supported in
DSL'.

The limitation of this approach is mainly that we cannot override a type
if the method already contains a `reifyAs` annotation. We accept this
limitation since (presumably) the standard and third-party libraries
won't have the annotations in the first place.

This commit does not provide a solution for method overloading.